Simple function to tidy Strategic Analytics for Improvement and Learning (SAIL) datasets from www.va.gov

Simple function to tidy Strategic Analytics for Improvement and Learning (SAIL) datasets from www.va.gov

sail_tidy(url)

Arguments

url
a SAIL dataset url from: https://www.va.gov/QUALITYOFCARE/measure-up/Strategic_Analytics_for_Improvement_and_Learning_SAIL.asp

Value

a [tibble][tibble::tibble-package] with SAIL metrics in tidy format

Examples

if (FALSE) {
url <- "https://www.va.gov/QUALITYOFCARE/SAIL_FY19_Q1/SAIL-2_albany_fy19q1.xlsx"
sail_tidy(url)

sail_tidy(
     "https://www.va.gov/QUALITYOFCARE/SAIL_FY19_Q1/SAIL-5_clarksburg_fy19q1.xlsx")
}